Fixed some warnings

+++ src/main.cc
@@ -18,7 +18,7 @@
 #include <cli.h>
 
 inline unsigned long long micro_sec() {
-  return std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
+    return static_cast<unsigned long long>(std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count());
 }
 
 int main(int argc, char **argv) {
